+++ addOns/websocket/CHANGELOG.md
@@ -4,7 +4,8 @@ All notable changes to this add-on will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## Unreleased
-
+### Changed
+- Update for Fuzzer payload replacement strategy changes (Cluster Bomb and Pitchfork).
 
 ## [37] - 2026-05-07
 ### Changed